Moves rehype-plugins into folder
